requirements.txt
cycler==0.10.0
decorator==4.3.0
kiwisolver==1.0.1
matplotlib==2.2.2
networkx==2.1
numpy==1.14.3
pandas==0.22.0
Pillow==5.1.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
scikit-image==0.13.1
scipy==1.1.0
six==1.11.0
torch==0.3.1
torchvision==0.2.1
tqdm==4.23.3